Fix migratin to SQLite compatible

+++ database/migrations/2020_09_02_105413_rename_portfolio_tables.php
@@ -13,15 +13,17 @@ class RenamePortfolioTables extends Migration
      */
     public function up()
     {
-        Schema::table('work_categories', function (Blueprint $table) {
-            $table->dropForeign(['work_id']);
-            $table->dropForeign(['category_id']);
-        });
+        if (config('database.default') == 'mysql') {
+            Schema::table('work_categories', function (Blueprint $table) {
+                $table->dropForeign(['work_id']);
+                $table->dropForeign(['category_id']);
+            });
 
-        Schema::table('work_tags', function (Blueprint $table) {
-            $table->dropForeign(['work_id']);
-            $table->dropForeign(['tag_id']);
-        });
+            Schema::table('work_tags', function (Blueprint $table) {
+                $table->dropForeign(['work_id']);
+                $table->dropForeign(['tag_id']);
+            });
+        }
 
         Schema::rename('works', 'portfolio_works');
         Schema::rename('categories', 'portfolio_categories');
@@ -29,19 +31,43 @@ class RenamePortfolioTables extends Migration
         Schema::rename('work_categories', 'portfolio_work_categories');
         Schema::rename('work_tags', 'portfolio_work_tags');
 
-        Schema::table('portfolio_work_categories', function (Blueprint $table) {
-            $table->renameColumn('work_id', 'portfolio_work_id');
-            $table->foreign('portfolio_work_id')->references('id')->on('portfolio_works')->onDelete('cascade')->onUpdate('cascade');
-            $table->renameColumn('category_id', 'portfolio_category_id');
-            $table->foreign('portfolio_category_id')->references('id')->on('portfolio_categories')->onDelete('cascade')->onUpdate('cascade');
-        });
+        if (config('database.default') == 'mysql') {
+            Schema::table('portfolio_work_categories', function (Blueprint $table) {
+                $table->renameColumn('work_id', 'portfolio_work_id');
+                $table->foreign('portfolio_work_id')->references('id')->on('portfolio_works')->onDelete('cascade')->onUpdate('cascade');
+                $table->renameColumn('category_id', 'portfolio_category_id');
+                $table->foreign('portfolio_category_id')->references('id')->on('portfolio_categories')->onDelete('cascade')->onUpdate('cascade');
+            });
 
-        Schema::table('portfolio_work_tags', function (Blueprint $table) {
-            $table->renameColumn('work_id', 'portfolio_work_id');
-            $table->foreign('portfolio_work_id')->references('id')->on('portfolio_works')->onDelete('cascade')->onUpdate('cascade');
-            $table->renameColumn('tag_id', 'portfolio_tag_id');
-            $table->foreign('portfolio_tag_id')->references('id')->on('portfolio_tags')->onDelete('cascade')->onUpdate('cascade');
-        });
+            Schema::table('portfolio_work_tags', function (Blueprint $table) {
+                $table->renameColumn('work_id', 'portfolio_work_id');
+                $table->foreign('portfolio_work_id')->references('id')->on('portfolio_works')->onDelete('cascade')->onUpdate('cascade');
+                $table->renameColumn('tag_id', 'portfolio_tag_id');
+                $table->foreign('portfolio_tag_id')->references('id')->on('portfolio_tags')->onDelete('cascade')->onUpdate('cascade');
+            });
+        }
+
+        if (config('database.default') == 'sqlite') {
+            Schema::dropIfExists('portfolio_work_categories');
+            Schema::create('portfolio_work_categories', function (Blueprint $table) {
+                $table->uuid('id')->primary();
+                $table->uuid('portfolio_work_id')->index();
+                $table->foreign('portfolio_work_id')->references('id')->on('portfolio_works')->onDelete('cascade')->onUpdate('cascade');
+                $table->uuid('portfolio_category_id')->index();
+                $table->foreign('portfolio_category_id')->references('id')->on('portfolio_categories')->onDelete('cascade')->onUpdate('cascade');
+                $table->timestamps();
+            });
+
+            Schema::dropIfExists('portfolio_work_tags');
+            Schema::create('portfolio_work_tags', function (Blueprint $table) {
+                $table->uuid('id')->primary();
+                $table->uuid('portfolio_work_id')->index();
+                $table->foreign('portfolio_work_id')->references('id')->on('portfolio_works')->onDelete('cascade')->onUpdate('cascade');
+                $table->uuid('portfolio_tag_id')->index();
+                $table->foreign('portfolio_tag_id')->references('id')->on('portfolio_tags')->onDelete('cascade')->onUpdate('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -51,15 +77,17 @@ class RenamePortfolioTables extends Migration
      */
     public function down()
     {
-        Schema::table('portfolio_work_categories', function (Blueprint $table) {
-            $table->dropForeign(['portfolio_work_id']);
-            $table->dropForeign(['portfolio_category_id']);
-        });
+        if (config('database.default') == 'mysql') {
+            Schema::table('portfolio_work_categories', function (Blueprint $table) {
+                $table->dropForeign(['portfolio_work_id']);
+                $table->dropForeign(['portfolio_category_id']);
+            });
 
-        Schema::table('portfolio_work_tags', function (Blueprint $table) {
-            $table->dropForeign(['portfolio_work_id']);
-            $table->dropForeign(['portfolio_tag_id']);
-        });
+            Schema::table('portfolio_work_tags', function (Blueprint $table) {
+                $table->dropForeign(['portfolio_work_id']);
+                $table->dropForeign(['portfolio_tag_id']);
+            });
+        }
 
         Schema::rename('portfolio_work_tags', 'work_tags');
         Schema::rename('portfolio_work_categories', 'work_categories');
@@ -67,18 +95,20 @@ class RenamePortfolioTables extends Migration
         Schema::rename('portfolio_categories' ,'categories');
         Schema::rename('portfolio_works' ,'works');
 
-        Schema::table('work_categories', function (Blueprint $table) {
-            $table->renameColumn('portfolio_work_id' ,'work_id');
-            $table->foreign('work_id')->references('id')->on('works')->onDelete('cascade')->onUpdate('cascade');
-            $table->renameColumn('portfolio_category_id', 'category_id');
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade')->onUpdate('cascade');
-        });
+        if (config('database.default') == 'mysql') {
+            Schema::table('work_categories', function (Blueprint $table) {
+                $table->renameColumn('portfolio_work_id' ,'work_id');
+                $table->foreign('work_id')->references('id')->on('works')->onDelete('cascade')->onUpdate('cascade');
+                $table->renameColumn('portfolio_category_id', 'category_id');
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade')->onUpdate('cascade');
+            });
 
-        Schema::table('work_tags', function (Blueprint $table) {
-            $table->renameColumn('portfolio_work_id', 'work_id');
-            $table->foreign('work_id')->references('id')->on('works')->onDelete('cascade')->onUpdate('cascade');
-            $table->renameColumn('portfolio_tag_id', 'tag_id');
-            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade')->onUpdate('cascade');
-        });
+            Schema::table('work_tags', function (Blueprint $table) {
+                $table->renameColumn('portfolio_work_id', 'work_id');
+                $table->foreign('work_id')->references('id')->on('works')->onDelete('cascade')->onUpdate('cascade');
+                $table->renameColumn('portfolio_tag_id', 'tag_id');
+                $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade')->onUpdate('cascade');
+            });
+        }
     }
 }
